simplified ui improved renderer size and capabilities and started new item system

+++ doomtst/renderer/HandleMap.h
@@ -166,7 +166,7 @@ namespace handle {
 					return free_ids.pop();
 				}
 				else {
-					return elems.length();
+					return static_cast<uint32_t>(elems.length());
 				}
 			}
